examples/rl/pyproject.toml
[project]
name = "rl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio>=4.0.0",
    "datasets>=4.0.0",
    "httpx>=0.28.1",
    "openai>=1.99.6",
    "openpipe-art[backend]>=0.4.4",
    "pydantic>=2.11.7",
    "rich>=14.1.0",
    "tenacity>=9.1.2",
    "transformers<4.54.0",
    "trl==0.19.1",
]

